Refactored to introduce base interfaces for the supported message types. This will allow someone to extend the RFC3164 and RFC5424 messages.
